[Config] Use endpoint url for accessing paddle and opensearch
The configuration files include the host name and port details,
which are used to construct the endpoint URLs. This approach adds
unnecessary code complexity and risks incorrect URL formation,
especially when the server uses different security protocols.

To avoid this, update the configuration file to allow users
to directly specify the expected URLs.

Signed-off-by: Vaibhav Mahajan <vamahaja@redhat.com,vaibhav.sukhadev.mahajan@ibm.com>